+++ src/main/java/org/attias/open/interactive/simulation/core/backend/engine/AppConfiguration.java
@@ -9,6 +9,7 @@ import java.io.InputStream;
 /**
  * For specific platform to run on the engine
  */
+@SuppressWarnings("unused")
 public class AppConfiguration {
     // The environment variable used to tell the runner where the OIS project jar exists.
     // It will load the jar to run with the engine
@@ -18,6 +19,12 @@ public class AppConfiguration {
     public static final String ENV_PROJECT_ASSETS_DIR = "OIS_ENV_PROJECT_ASSETS_DIR";
     // The environment variable used to tell the runner what is the name of the artifact to generate on publish
     public static final String ENV_PROJECT_NAME = "OIS_ENV_PROJECT_NAME";
+    // The environment variable used to tell the runner what is the group name of the artifact to generate on publish
+    public static final String ENV_PROJECT_GROUP = "OIS_ENV_PROJECT_GROUP";
+    // The environment variable used to tell the runner what is the version of the artifact to generate on publish
+    public static final String ENV_PROJECT_VERSION = "OIS_ENV_PROJECT_VERSION";
+    // The environment variable used to tell the runner what is the version number (typically incremented with each release) of the artifact to generate on publish
+    public static final String ENV_PROJECT_VERSION_NUMBER = "OIS_ENV_PROJECT_VERSION_NUM";
     // The environment variable used when running not on publish, to tell the runner where the android sdk
     // If not exists, it will search the value from environment variable named 'ANDROID_HOME';
     public static final String ENV_ANDROID_SDK_PATH = "OIS_ENV_ANDROID_SDK_PATH";
@@ -50,8 +57,6 @@ public class AppConfiguration {
     }
 
     public static AppConfiguration getRunnerConfigurations(InputStream assetsProjectConfiguration) throws IOException {
-        AppConfiguration configuration = new AppConfiguration();
-        configuration.setProjectConfiguration(ProjectUtils.getProjectConfiguration(assetsProjectConfiguration));
-        return configuration;
+        return new AppConfiguration().setProjectConfiguration(ProjectUtils.getProjectConfiguration(assetsProjectConfiguration));
     }
 }
