+++ src/main/java/org/attias/open/interactive/simulation/core/utils/XmlUtils.java
@@ -6,7 +6,6 @@ import org.xml.sax.SAXException;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerConfigurationException;
 import javax.xml.transform.TransformerException;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
@@ -14,6 +13,7 @@ import javax.xml.transform.stream.StreamResult;
 import java.io.IOException;
 import java.nio.file.Path;
 
+@SuppressWarnings("unused")
 public class XmlUtils {
     private static final DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
     private static final TransformerFactory transformerFactory = TransformerFactory.newInstance();
